# [MXNET-249] Add inplace support to mkldnn sum