--- a/src/operator/nn/mkldnn/mkldnn_sum.cc
+++ b/src/operator/nn/mkldnn/mkldnn_sum.cc
@@ -65,22 +65,24 @@ void MKLDNNSumForward(const nnvm::NodeAttrs& attrs, const OpContext &ctx,
     in_pds[i] = in_mem->get_primitive_desc();
     pd_same = pd_same && (in_pds[i] == in_pds[0]);
   }
-  mkldnn::sum::primitive_desc pdesc(scales, in_pds);
 
+  mkldnn::sum::primitive_desc pdesc(scales, in_pds);
+  pd_same = pd_same && (pdesc.dst_primitive_desc() == in_pds[0]);
+  auto out_mem = const_cast<NDArray&>(out_data).CreateMKLDNNData(pdesc.dst_primitive_desc());
+  bool addr_same = out_mem->get_data_handle() == inputs[0].GetMKLDNNData()->get_data_handle();
   if ((req == kWriteTo) ||
-      (req == kWriteInplace && pd_same && (pdesc.dst_primitive_desc() == in_pds[0]))) {
+      (req == kWriteInplace && pd_same && addr_same)) {
     // do sum computation directly on output NDArray
-    auto out_mem = const_cast<NDArray&>(out_data).CreateMKLDNNData(pdesc.dst_primitive_desc());
     MKLDNNStream *stream = MKLDNNStream::Get();
     stream->RegisterPrim(mkldnn::sum(pdesc, in_prims, *out_mem));
     stream->Submit();
   } else {
     // req == kWriteInplace but cannot be handled by mkldnn and
     // req == kAddTo will run into this branch
-    auto out_mem = CreateMKLDNNMem(out_data, pdesc.dst_primitive_desc(), req);
+    auto mem = CreateMKLDNNMem(out_data, pdesc.dst_primitive_desc(), req);
     MKLDNNStream *stream = MKLDNNStream::Get();
-    stream->RegisterPrim(mkldnn::sum(pdesc, in_prims, *out_mem.second));
-    CommitOutput(out_data, out_mem);
+    stream->RegisterPrim(mkldnn::sum(pdesc, in_prims, *mem.second));
+    CommitOutput(out_data, mem);
     stream->Submit();
   }
 }