@TaoLv 2018-03-29
FYI, below is the inference perfermance of resnet50 on my local server:

BS | inplace-sum | master (0ff490b) | Speedup
:--: | :--: | :--: | :--:
1 | 85.11 | 82.02 | 103.78%
2 | 122.72 | 117.91 | 104.08%
4 | 152.03 | 146.68 | 103.65%
8 | 191.20 | 181.78 | 105.18%
16 | 214.48 | 198.97 | 107.79%
32 | 222.46 | 202.50 | 109.86%
64 | 237.22 | 203.89 | 116.35%
128 | 230.75 | 197.16 | 117.03%