@zheng-da 2018-03-29
There is another place.
https://github.com/apache/incubator-mxnet/blob/master/src/operator/nn/mkldnn/mkldnn_base.cc#L122
We should also avoid memory copy here if possible.