@TaoLv 2018-03-29
Yes, same behavior for these two conditions. Just for better code readability here. Conditional check will be a little complicated if try to merge these two branches together. Any good ideas for this?


On 29 Mar 2018, at 11:57 PM, Ashok Emani <notifications@github.com<mailto:notifications@github.com>> wrote:


@ashokei commented on this pull request.

________________________________

In src/operator/nn/mkldnn/mkldnn_sum.cc<https://github.com/apache/incubator-mxnet/pull/10315#discussion_r178102539>:

>    }
   mkldnn::sum::primitive_desc pdesc(scales, in_pds);

-  auto out_mem = CreateMKLDNNMem(out_data, pdesc.dst_primitive_desc(), req);
-  MKLDNNStream *stream = MKLDNNStream::Get();
-  stream->RegisterPrim(mkldnn::sum(pdesc, in_prims, *out_mem.second));
-  CommitOutput(out_data, out_mem);
-  stream->Submit();
+  if (req == kWriteTo) {


seems same as inplace option below, what am i missing.

—
You are receiving this because you authored the thread.
Reply to this email directly, view it on GitHub<https://github.com/apache/incubator-mxnet/pull/10315#pullrequestreview-108105980>, or mute the thread<https://github.com/notifications/unsubscribe-auth/AVZehp0VLcoAXYCyfgzuEqT1_4zTMZj2ks5tjQRWgaJpZM4TAS9w>.