@piiswrong 2018-03-29
You can merge these. Also looks like you didn't handle addto?