@TaoLv 2018-03-30
@piiswrong `kAddTo` is handled in the else branch by `CreateMKLDNNMem` and `CommitOutput`. I added some comments to describe that.
Another restriction for mkldnn inplace sum: output should have same memory address with the first input. That means mkldnn sum operator can only perform inplace computation on the first input.